refactoring relational make relational, range_restrict and linear associated methods move range restriction and linearization algorithms to their own modules
