+++ razor-fol/src/transform.rs
@@ -2,9 +2,11 @@
 mod cnf;
 mod dnf;
 mod gnf;
+mod linear;
 mod nnf;
 mod pnf;
-pub mod relationalize;
+mod range_restrict;
+mod relational;
 mod simplify;
 mod snf;
 use thiserror::Error;
@@ -12,9 +14,10 @@ use thiserror::Error;
 pub use cnf::CNF;
 pub use dnf::DNF;
 pub use gnf::{PcfSet, GNF, PCF};
+pub use linear::LinearConfig;
 pub use nnf::NNF;
 pub use pnf::PNF;
-pub use relationalize::{RelClause, Relational, Variable};
+pub use relational::{RelClause, Relational, RelationalConfig, Variable};
 pub use snf::SNF;
 
 /// Is the type of errors returned by syntactic transformations.
