+++ razor-fol/src/transform/relational.rs
@@ -160,8 +160,7 @@ impl From<RelClause> for FOF {
 /// transform positive formulae in the head and body of a [`GNF`] to relational forms.
 ///
 /// **Hint**: A relationalized formula contains only (flat) variable terms with
-/// no function symbol nor constants. [`Relationalizer`] transforms suitable formulae
-/// to a relational form.
+/// no function symbol nor constants.
 ///
 /// [`GNF`]: crate::transform::GNF
 #[derive(Clone, Debug)]
@@ -255,8 +254,8 @@ impl std::fmt::Display for Relational {
 
 /// Provides the relationalization algorithm through the [`transform`] method.
 ///
-/// [`transform`]: Relationalizer::transform()
-pub struct Relationalizer {
+/// [`transform`]: Relational::transform()
+pub struct RelationalConfig {
     // Is the [`Generator`] instance used to generate variable names used as placeholders
     // in flattened formulae.
     flattening_generator: Generator,
@@ -272,7 +271,7 @@ pub struct Relationalizer {
     generated_variables: Vec<V>,
 }
 
-impl Relationalizer {
+impl RelationalConfig {
     /// Use `generator` to generate flattening variable names.
     pub fn set_flattening_generator(&mut self, generator: Generator) {
         self.flattening_generator = generator;
@@ -281,14 +280,6 @@ impl Relationalizer {
     /// Use `generator` to create function predicate names.
     pub fn set_predicate_generator(&mut self, generator: Generator) {
         self.function_predicate_generator = generator;
-    }
-
-    // Resets the Relationalizer, preparing for the next transformation.
-    fn reset(&mut self) {
-        self.constant_predicate_generator.reset();
-        self.function_predicate_generator.reset();
-        self.flattening_generator.reset();
-        self.generated_variables.clear();
     }
 
     // Generates a new variable using the `flattening_generator` and keeps track of it in
@@ -300,8 +291,8 @@ impl Relationalizer {
     }
 }
 
-impl Default for Relationalizer {
-    /// Creates a new `Relationalizer` instance with default generators and symbols:
+impl Default for RelationalConfig {
+    /// Creates a new `RelationalConfig` instance with default generators and symbols:
     ///   * The equality symbol is `=`.
     ///   * Variables introduced by flattening are prefixed with `?`.
     ///   * Function predicates are prefixed with `$`.    
@@ -317,7 +308,7 @@ impl Default for Relationalizer {
 
 // Recursively flattens a term and returns the resulting formula together with a placeholder variable
 // for the term. Nothing needs to be done if the input term is a variable.
-fn flatten_term(term: &Complex, relationalizer: &mut Relationalizer) -> (Option<RelClause>, V) {
+fn flatten_term(term: &Complex, relationalizer: &mut RelationalConfig) -> (Option<RelClause>, V) {
     match term {
         Complex::Var { variable } => (None, variable.clone()),
         Complex::Const { constant } => {
@@ -362,7 +353,7 @@ fn flatten_term(term: &Complex, relationalizer: &mut Relationalizer) -> (Option<
 }
 
 // Applies top level flattening on the input formula.
-fn flatten_formula(clause_set: &PcfSet, relationalizer: &mut Relationalizer) -> Relational {
+fn flatten_formula(clause_set: &PcfSet, relationalizer: &mut RelationalConfig) -> Relational {
     clause_set
         .iter()
         .map(|clause| {
@@ -422,7 +413,7 @@ fn flatten_formula(clause_set: &PcfSet, relationalizer: &mut Relationalizer) -> 
 fn equation_rewrites<'a>(
     clause_set: &'a Relational,
     map: &mut HashMap<&'a V, &'a V>,
-    relationalizer: &mut Relationalizer,
+    relationalizer: &mut RelationalConfig,
 ) {
     clause_set.iter().for_each(|clause| {
         clause.iter().for_each(|atomic| match atomic {
@@ -459,7 +450,10 @@ fn equation_rewrites<'a>(
 
 // Simplify tries to minimize the use of existential variables (generated by relationalize) by replacing
 // them with universal or other existential variables that are syntactically equal with them.
-fn simplify_equations(clause_set: &Relational, relationalizer: &mut Relationalizer) -> Relational {
+fn simplify_equations(
+    clause_set: &Relational,
+    relationalizer: &mut RelationalConfig,
+) -> Relational {
     let mut map = HashMap::new();
     equation_rewrites(clause_set, &mut map, relationalizer);
 
@@ -471,11 +465,11 @@ fn simplify_equations(clause_set: &Relational, relationalizer: &mut Relationaliz
 }
 
 impl PcfSet {
-    /// Is similar to [`PcfSet::relational`] but uses a custom [`Relationalizer`].
+    /// Is similar to [`PcfSet::relational`] but uses a custom [`RelationalConfig`].
     /// **Example**:
     /// ```rust
     /// # use razor_fol::syntax::FOF;
-    /// use razor_fol::transform::relationalize::Relationalizer;
+    /// use razor_fol::transform::RelationalConfig;
     ///
     /// let fof = "P(x) -> P(f(x)) & Q('c)".parse::<FOF>().unwrap();
     /// let gnfs = fof.gnf();
@@ -486,13 +480,24 @@ impl PcfSet {
     ///     transformed.to_string()
     /// );
     /// ```
-    pub fn relational_with(&self, relationalizer: &mut Relationalizer) -> Relational {
+    pub fn relational_with(&self, relationalizer: &mut RelationalConfig) -> Relational {
         let flattened = flatten_formula(self, relationalizer);
         let simplified = simplify_equations(&flattened, relationalizer);
-        let result = remove_reflexive_equations(simplified);
-        relationalizer.reset(); // prepare for next call to `transform`
-
-        result
+        simplified
+            .into_clauses()
+            .into_iter()
+            .map(|clause| {
+                clause
+                    .into_atomics()
+                    .into_iter()
+                    // remove reflexive equations:
+                    .filter(|atomic| match atomic {
+                        Atomic::Equals(this) => this.left != this.right,
+                        _ => true,
+                    })
+                    .into()
+            })
+            .into()
     }
 
     /// Applies the relationalization algorithm on the receiver and returns a relational formula.    
@@ -517,7 +522,7 @@ impl PcfSet {
     /// **Example**:
     /// ```rust
     /// # use razor_fol::syntax::FOF;
-    /// use razor_fol::transform::relationalize::Relationalizer;
+    /// use razor_fol::transform::RelationalConfig;
     ///
     /// let fof = "P(x) -> P(f(x)) & Q('c)".parse::<FOF>().unwrap();
     /// let gnfs = fof.gnf();
@@ -529,255 +534,14 @@ impl PcfSet {
     /// );
     /// ```
     pub fn relational(&self) -> Relational {
-        self.relational_with(&mut Relationalizer::default())
+        self.relational_with(&mut RelationalConfig::default())
     }
 }
 
-/// Is used to expand implicit equations by replacing variables that appear in more than
-/// one position of a formula with freshly generated variables. The expansion algorithm
-/// is provided by the [`transform`] method.
-///
-/// [`transform`]: EqualityExpander::transform()
-pub struct EqualityExpander {
-    // Is the symbol used to convert equality to a predicate.
-    equality_symbol: String,
-
-    // Is the [`Generator`] instance used to generate new variable names when variables
-    // appear in more than one position.
-    equality_generator: Generator,
-}
-
-impl EqualityExpander {
-    /// Use `symbol` for equality predicates.
-    pub fn set_equality_symbol<S: Into<String>>(&mut self, symbol: S) {
-        self.equality_symbol = symbol.into();
-    }
-
-    /// Use `generator` to distinguish variables that appear in more than one positions.
-    pub fn set_equality_generator(&mut self, generator: Generator) {
-        self.equality_generator = generator;
-    }
-
-    fn helper(&self, rel: &Relational) -> Relational {
-        let mut vars = HashMap::<V, i32>::new();
-        rel.iter()
-            .map(|clause| {
-                clause
-                    .iter()
-                    .flat_map(|atomic| match atomic {
-                        Atomic::Atom(this) => {
-                            let mut equations = Vec::<Atomic<_>>::new();
-                            let mut new_terms = Vec::new();
-                            for variable in &this.terms {
-                                vars.entry(variable.symbol().clone())
-                                    .and_modify(|count| {
-                                        let new_var = V(self
-                                            .equality_generator
-                                            .indexed_symbol(variable.0.to_string(), *count));
-
-                                        let left = variable.clone();
-                                        let right = new_var.clone().into();
-
-                                        let eq = Atom {
-                                            predicate: self.equality_symbol.as_str().into(),
-                                            terms: vec![left, right],
-                                        }
-                                        .into();
-                                        equations.push(eq);
-                                        new_terms.push(new_var.into());
-                                        *count += 1;
-                                    })
-                                    .or_insert_with(|| {
-                                        new_terms.push(variable.clone().into());
-                                        0
-                                    });
-                            }
-                            equations.push(
-                                Atom {
-                                    predicate: this.predicate.clone(),
-                                    terms: new_terms,
-                                }
-                                .into(),
-                            );
-                            equations
-                        }
-                        Atomic::Equals(this) => {
-                            let mut equations = Vec::<Atomic<_>>::new();
-                            let left = &this.left;
-                            let right = &this.right;
-                            let mut new_left = V("".into());
-                            let mut new_right = V("".into());
-
-                            vars.entry(left.symbol().clone())
-                                .and_modify(|count| {
-                                    let new_var = V(self
-                                        .equality_generator
-                                        .indexed_symbol(left.to_string(), *count));
-
-                                    let left = left.clone();
-                                    let right = new_var.clone().into();
-
-                                    let eq = Atom {
-                                        predicate: self.equality_symbol.as_str().into(),
-                                        terms: vec![left, right],
-                                    }
-                                    .into();
-                                    equations.push(eq);
-                                    new_left = new_var.into();
-                                    *count += 1;
-                                })
-                                .or_insert_with(|| {
-                                    new_left = left.symbol().clone();
-                                    0
-                                });
-
-                            vars.entry(right.symbol().clone())
-                                .and_modify(|count| {
-                                    let new_var = V(self
-                                        .equality_generator
-                                        .indexed_symbol(right.to_string(), *count));
-
-                                    let left = right.clone();
-                                    let right = new_var.clone().into();
-
-                                    let eq = Atom {
-                                        predicate: self.equality_symbol.as_str().into(),
-                                        terms: vec![left, right],
-                                    }
-                                    .into();
-                                    equations.push(eq);
-                                    new_right = new_var.into();
-                                    *count += 1;
-                                })
-                                .or_insert_with(|| {
-                                    new_right = right.symbol().clone();
-                                    0
-                                });
-
-                            equations.push(
-                                Equals {
-                                    left: new_left.into(),
-                                    right: new_right.into(),
-                                }
-                                .into(),
-                            );
-                            equations
-                        }
-                    })
-                    .into()
-            })
-            .into()
-    }
-
-    /// Replaces replaces any varialbe `v` that appears in more than one position of `formula`
-    /// with a fresh variable `y` and an atom `=(v, y)` is conjoined with `formula`.
-    ///
-    /// **Example**:
-    /// ```rust
-    /// # use razor_fol::syntax::FOF;
-    /// use razor_fol::transform::relationalize::{Relationalizer, EqualityExpander};
-    ///
-    /// let fof = "P(x) -> P(f(x)) & Q('c)".parse::<FOF>().unwrap();
-    /// let gnfs = fof.gnf();
-    /// let gnf_head = gnfs[0].head();
-    /// let relational = gnf_head.relational();
-    /// let transformed = EqualityExpander::default().transform(&relational);
-    /// assert_eq!(    
-    ///     r"(((($f(x, ?0) ∧ =(?0, ~?0:0)) ∧ P(~?0:0)) ∧ @c(?1)) ∧ =(?1, ~?1:0)) ∧ Q(~?1:0)",
-    ///     transformed.to_string()
-    /// );
-    /// ```
-    pub fn transform(&self, rel: &Relational) -> Relational {
-        let transformed = self.helper(rel);
-        let result = remove_reflexive_equations(transformed);
-        result
-    }
-}
-
-impl Default for EqualityExpander {
-    /// Creates a new `EqualityExpander` instance with default generators and symbols:
-    /// * The equality symbol is `=`.
-    /// * Variables appearing in more than one position are distinguished with `~` as the
-    /// prefix, followed by `:` and a unique index.
-    fn default() -> Self {
-        Self {
-            equality_symbol: "=".into(),
-            equality_generator: Generator::new().set_prefix("~").set_delimiter(":"),
-        }
-    }
-}
-
-/// Given a `formula`, and a list of variables `range`, ensures that every
-/// variable in `range` appears at least once in `formula`. This is done by conjoining atomic
-/// formulae in the form of `RR(v)` where `RR` is a "range restriction" predicate with
-/// `symbol` as the predicate symbol. The transformation fails if the `formula` is not
-/// relationalized.
-/// The underlying algorithm assumes that the input is negation and quantifier-free;
-/// that is, `¬`, `→`, `⇔`, `∃`, `∀` are not allowed as connectives.
-///
-/// **Note**: the term "range restriction" is borrowed from the database literature.
-///
-/// **Example**:
-/// ```rust
-/// use razor_fol::syntax::FOF;
-/// use razor_fol::v;
-/// use razor_fol::transform::relationalize::{Relationalizer, range_restrict};
-///
-/// let fof = "R(x, y) -> P(x) & Q(y)".parse::<FOF>().unwrap();
-/// let gnfs = fof.gnf();
-/// let gnf_head = gnfs[0].head();
-/// let relational = gnf_head.relational();
-/// let transformed = range_restrict(&relational, &vec![v!(x), v!(z)], "RR");
-/// assert_eq!(r"(P(x) ∧ Q(y)) ∧ RR(z)", transformed.to_string());
-/// ```
-pub fn range_restrict(rel: &Relational, range: &[V], symbol: &str) -> Relational {
-    rel.iter()
-        .map(|clause| {
-            let free = clause.free_vars();
-            let mut range = Vec::from(range);
-            range.retain(|x| !free.contains(&x));
-            let mut atomics = clause.atomics().to_vec();
-            atomics.extend(rr_helper(&range, symbol).into_atomics());
-            atomics.into()
-        })
-        .into()
-}
-
-// Is a helper for `range_restrict` to build range_restriction conjuncts.
-#[inline(always)]
-fn rr_helper(range: &[V], symbol: &str) -> RelClause {
-    let mut result = Vec::new();
-    for v in range {
-        result.push(
-            Atom {
-                predicate: symbol.into(),
-                terms: vec![v.clone().into()],
-            }
-            .into(),
-        );
-    }
-    result.into()
-}
-
-fn remove_reflexive_equations(rel: Relational) -> Relational {
-    rel.into_clauses()
-        .into_iter()
-        .map(|clause| {
-            clause
-                .into_atomics()
-                .into_iter()
-                .filter(|atomic| match atomic {
-                    Atomic::Equals(this) => this.left != this.right,
-                    _ => true,
-                })
-                .into()
-        })
-        .into()
-}
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{fof, transform::PCF, v};
+    use crate::{fof, transform::PCF};
 
     // Assumes the input in GNF
     fn clause_set(fof: FOF) -> Vec<PcfSet> {
@@ -790,7 +554,7 @@ mod tests {
     fn flatten(fof: FOF) -> String {
         let rels = clause_set(fof)
             .iter()
-            .map(|f| flatten_formula(f, &mut Relationalizer::default()))
+            .map(|f| flatten_formula(f, &mut RelationalConfig::default()))
             .map(FOF::from)
             .collect::<Vec<_>>();
         format!("{:?}", rels)
@@ -869,107 +633,6 @@ mod tests {
             "[(@c(?0) & P(?0)) | (@d(?1) & Q(?1))]",
             flatten(fof!((P(@c)) | (Q(@d)))),
         };
-    }
-
-    fn linear(fof: FOF) -> String {
-        let rels = clause_set(fof)
-            .iter()
-            .map(|f| flatten_formula(f, &mut Relationalizer::default()))
-            .map(|f| EqualityExpander::default().helper(&f))
-            .map(FOF::from)
-            .collect::<Vec<_>>();
-        format!("{:?}", rels)
-    }
-
-    #[test]
-    fn test_expand_equality() {
-        // Note: the body of sequents in the tests below are needed to satisfy
-        // the requirement that heads of geometric sequents with no free variables
-        // in their head gets compressed; otherwise, these tests wouldn't not
-        // be interesting.
-        assert_eq!("[]", linear(fof!('|')));
-        assert_eq!("[false]", linear(fof!(_|_)));
-        assert_eq!("[P()]", linear(fof!(P())));
-        assert_eq!("[P(x, y)]", linear(fof!(P(x, y))));
-        assert_eq!("[=(x, ~x:0) & P(x, ~x:0)]", linear(fof!(P(x, x))));
-        assert_eq!(
-            "[(P(x, y) & =(x, ~x:0)) & Q(~x:0)]",
-            linear(fof!({P(x, y)} -> {[P(x, y)] & [Q(x)]}))
-        );
-        assert_eq!(
-            "[((P(x, y) & =(x, ~x:0)) & =(y, ~y:0)) & Q(~x:0, ~y:0)]",
-            linear(fof!({P(x, y)} -> {[P(x, y)] & [Q(x, y)]}))
-        );
-        assert_eq!(
-            "[((P(x) & =(x, ~x:0)) & =(y, ~y:0)) & Q(y, ~x:0, ~y:0)]",
-            linear(fof!({P(x, y)} -> {[P(x)] & [Q(y, x, y)]}))
-        );
-        assert_eq!(
-            "[(((P(x) & =(x, ~x:0)) & Q(~x:0)) & =(x, ~x:1)) & R(~x:1)]",
-            linear(fof!([P(x)] -> [{ [P(x)] & [Q(x)] } & { R(x) }]))
-        );
-        assert_eq!(
-            "[P(x, y) | (=(x, ~x:0) & Q(~x:0))]",
-            linear(fof!({P(x, y)} -> {[P(x, y)] | [Q(x)]}))
-        );
-        assert_eq!(
-            "[P(x, y) | ((=(x, ~x:0) & =(y, ~y:0)) & Q(~x:0, ~y:0))]",
-            linear(fof!({P(x, y)} -> {[P(x, y)] | [Q(x, y)]}))
-        );
-        assert_eq!(
-            "[P(x) | ((=(x, ~x:0) & =(y, ~y:0)) & Q(y, ~x:0, ~y:0))]",
-            linear(fof!({P(x, y)} -> {[P(x)] | [Q(y, x, y)]}))
-        );
-        assert_eq!(
-            "[(P(x) | (=(x, ~x:0) & Q(~x:0))) | (=(x, ~x:1) & R(~x:1))]",
-            linear(fof!([P(x)] -> [{ [P(x)] | [Q(x)] } | { R(x) }]))
-        );
-    }
-
-    fn rr(fof: FOF, range: &[V]) -> String {
-        let rels = clause_set(fof)
-            .iter()
-            .map(|f| flatten_formula(f, &mut Relationalizer::default()))
-            .map(|f| range_restrict(&f, range, "RR"))
-            .map(FOF::from)
-            .collect::<Vec<_>>();
-        format!("{:?}", rels)
-    }
-
-    #[test]
-    fn test_range_restrict() {
-        assert_eq!("[]", rr(fof!('|'), &vec![]));
-        assert_eq!("[]", rr(fof!('|'), &vec![v!(x), v!(y)]));
-        assert_eq!("[false]", rr(fof!(_|_), &vec![]));
-
-        assert_eq!("[P(x)]", rr(fof!(P(x)), &vec![]));
-        assert_eq!(
-            "[P(w, x, y) & RR(z)]",
-            rr(fof!(P(w, x, y)), &vec![v!(x), v!(y), v!(z)])
-        );
-
-        assert_eq!(
-            "[P(x) & Q(y)]",
-            rr(fof!({R(x, y)} -> {[P(x)] & [Q(y)]}), &vec![])
-        );
-        assert_eq!(
-            "[((P(v, w) & Q(x)) & RR(y)) & RR(z)]",
-            rr(
-                fof!({R(v, w, x)} -> {[P(v, w)] & [Q(x)]}),
-                &vec![v!(v), v!(w), v!(x), v!(y), v!(z)],
-            )
-            .to_string()
-        );
-
-        assert_eq!(
-            "[P(x) | Q(y)]",
-            rr(fof!({R(x, y)} -> {[P(x)] | [Q(y)]}), &vec![])
-        );
-
-        assert_eq!(
-            "[(P(x) & RR(y)) | (Q(y) & RR(x))]",
-            rr(fof!([P(x)] | [Q(y)]), &vec![v!(x), v!(y)])
-        );
     }
 
     #[test]
